move title hook to legacy slice (for consistancy)
