+++ src/features/profile/Profile.tsx
@@ -2,7 +2,8 @@ import { FC, useMemo } from 'react';
 import { Link, useLocation, useParams } from 'react-router-dom';
 import { Button } from '../../common/components';
 import NarrativeList from '../../common/components/NarrativeList/NarrativeList';
-import { useAppSelector, usePageTitle } from '../../common/hooks';
+import { useAppSelector } from '../../common/hooks';
+import { usePageTitle } from '../layout/layoutSlice';
 import { authUsername } from '../auth/authSlice';
 import PageNotFound from '../layout/PageNotFound';
 import classes from './Profile.module.scss';
